nlp_service/requirements.txt
flask
python-multipart
python-dotenv
PyMuPDF==1.23.0
docx2txt
openai
requests
